supabase/edge_functions/face-recognition/requirements.txt
# Python dependencies for face recognition Edge Function
# Install these in the Supabase Edge Function environment

face-recognition==1.3.0
numpy>=1.21.0
Pillow>=9.0.0
supabase>=2.0.0

